feat(learning): 实现选课和查询学习资格功能
- 添加选课功能，支持免费课程和收费课程的选课操作
- 实现查询学习资格功能，根据课程状态返回不同的学习状态
- 新增 XcChooseCourseService接口及其实现类
- 修改相关的实体类和 Mapper 文件

+++ learn-online-learning/learn-online-learning-model/src/main/java/com/xia/learning/model/po/XcChooseCourse.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -61,7 +62,7 @@ public class XcChooseCourse implements Serializable {
      */
     private Integer validDays;
 
-    private Float coursePrice;
+    private BigDecimal coursePrice;
 
     /**
      * 选课状态
